requirements.txt
asgiref==3.8.1
attrs==25.4.0
Django==5.2.3
django-filter==25.2
djangorestframework==3.16.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
